# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "AlternativeOut": {
        "description": "یک پروژهٔ جایگزین همراه پاسخ رد — §7.5.",
        "properties": {
          "match_score": {
            "title": "Match Score",
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectSummaryOut"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonOut"
            },
            "title": "Reasons",
            "type": "array"
          }
        },
        "required": [
          "project",
          "match_score"
        ],
        "title": "AlternativeOut",
        "type": "object"
      },
      "ApplicationIn": {
        "additionalProperties": false,
        "properties": {
          "motivation": {
            "maxLength": 500,
            "minLength": 10,
            "title": "Motivation",
            "type": "string"
          },
          "role_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          }
        },
        "required": [
          "motivation"
        ],
        "title": "ApplicationIn",
        "type": "object"
      },
      "ApplicationOut": {
        "properties": {
          "applicant_id": {
            "format": "uuid",
            "title": "Applicant Id",
            "type": "string"
          },
          "applicant_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicant Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_breakdown": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Breakdown"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "motivation": {
            "title": "Motivation",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "project_title_fa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Title Fa"
          },
          "role_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          },
          "role_title_fa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Title Fa"
          },
          "status": {
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "WAITLISTED",
              "WITHDRAWN"
            ],
            "title": "Status",
            "type": "string"
          },
          "status_fa": {
            "title": "Status Fa",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "applicant_id",
          "motivation",
          "status",
          "status_fa",
          "created_at"
        ],
        "title": "ApplicationOut",
        "type": "object"
      },
...
      "DecisionIn": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "ACCEPTED",
              "REJECTED",
              "WAITLISTED"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "decision"
        ],
        "title": "DecisionIn",
        "type": "object"
      },
      "DecisionOut": {
        "properties": {
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/AlternativeOut"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "application": {
            "$ref": "#/components/schemas/ApplicationOut"
          }
        },
        "required": [
          "application"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
...
      "DownloadUrlOut": {
        "properties": {
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          }
        },
        "required": [
          "download_url",
          "expires_in",
          "original_name"
        ],
        "title": "DownloadUrlOut",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "قالب ثابت خطا — §5.1. فقط برای مستندسازی OpenAPI.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FileOut": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          },
          "purpose": {
            "enum": [
              "DELIVERABLE",
              "RESOURCE",
              "PROJECT_COVER",
              "AVATAR",
              "MESSAGE"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "scan_status": {
            "enum": [
              "PENDING",
              "CLEAN",
              "INFECTED",
              "SKIPPED"
            ],
            "title": "Scan Status",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "uploaded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded At"
          }
        },
        "required": [
          "id",
          "original_name",
          "content_type",
          "size_bytes",
          "purpose",
          "scan_status"
        ],
        "title": "FileOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReasonOut": {
        "description": "دلیل آمادهٔ نمایش — §8.10. متن را بک‌اند می‌سازد، نه کلاینت.",
        "properties": {
          "contribution": {
            "title": "Contribution",
            "type": "number"
          },
          "polarity": {
            "enum": [
              "POSITIVE",
              "WARNING"
            ],
            "title": "Polarity",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "polarity",
          "contribution",
          "text"
        ],
        "title": "ReasonOut",
        "type": "object"
      },
...
      "UploadUrlIn": {
        "additionalProperties": false,
        "properties": {
          "content_type": {
            "maxLength": 150,
            "minLength": 3,
            "title": "Content Type",
            "type": "string"
          },
          "original_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Original Name",
            "type": "string"
          },
          "purpose": {
            "enum": [
              "DELIVERABLE",
              "RESOURCE",
              "PROJECT_COVER",
              "AVATAR",
              "MESSAGE"
            ],
            "title": "Purpose",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "original_name",
          "content_type",
          "size_bytes",
          "purpose"
        ],
        "title": "UploadUrlIn",
        "type": "object"
      },
      "UploadUrlOut": {
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "file_id": {
            "format": "uuid",
            "title": "File Id",
            "type": "string"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "max_bytes": {
            "title": "Max Bytes",
            "type": "integer"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "file_id",
          "upload_url",
          "method",
          "headers",
          "expires_in",
          "max_bytes"
        ],
        "title": "UploadUrlOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nAPI پلتفرم نوآوری و یادگیری صابر.\n\nپاسخ‌ها مستقیماً بدنهٔ داده هستند و خطاها قالب ثابت\n`{\"error\": {\"code\", \"message\", \"details\", \"trace_id\"}}` دارند.\n",
    "title": "SILP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/applications/mine": {
      "get": {
        "operationId": "my_applications_api_v1_applications_mine_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationOut"
                  },
                  "title": "Response My Applications Api V1 Applications Mine Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "درخواست‌های من",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/applications/{application_id}": {
      "delete": {
        "operationId": "withdraw_application_api_v1_applications__application_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "انصراف از درخواست",
        "tags": [
          "projects"
        ]
      }
    },
    "/api/v1/applications/{application_id}/decide": {
      "post": {
        "description": "§7.5 — پاسخ رد **همیشه** با سه پروژهٔ جایگزین برمی‌گردد.\n\nجایگزین‌ها از موتور توصیه‌گر می‌آیند و همان‌جا در پاسخ می‌نشینند تا\nکلاینت مجبور نشود یک درخواست دیگر بزند؛ «رد شدم» و «مسیر بهتری پیدا\nکردم» نباید به تأخیر شبکه گره بخورد.",
        "operationId": "decide_application_api_v1_applications__application_id__decide_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PROJECT_CAPACITY_FULL"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "تصمیم دربارهٔ درخواست",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/files/upload-url": {
      "post": {
        "description": "§5.9 — کلاینت با این آدرس مستقیم روی S3 می‌نویسد.",
        "operationId": "upload_url_api_v1_files_upload_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadUrlIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "FILE_TOO_LARGE"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "CONTENT_TYPE_NOT_ALLOWED"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "دریافت URL آپلود مستقیم",
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/{file_id}": {
      "delete": {
        "operationId": "delete_file_api_v1_files__file_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "File Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "حذف نرم فایل",
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/{file_id}/complete": {
      "post": {
        "description": "حجم واقعی و Magic Number اینجا بررسی می‌شوند — §11.1.",
        "operationId": "complete_upload_api_v1_files__file_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "File Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "اعلام پایان آپلود",
        "tags": [
          "files"
        ]
      }
    },
    "/api/v1/files/{file_id}/download-url": {
      "get": {
        "description": "آدرس کوتاه‌عمر با `Content-Disposition: attachment`.\n\nدر M2 فقط آپلودکننده دانلود می‌کند. دسترسی هم‌تیمی‌ها به کتابخانهٔ\nفایل پروژه از راه `GET /projects/{id}/files` می‌آید؛ تا آن زمان\nنشتی وجود ندارد.",
        "operationId": "download_url_api_v1_files__file_id__download_url_get",
        "parameters": [
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "File Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadUrlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "URL دانلود موقت",
        "tags": [
          "files"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/applications": {
      "get": {
        "description": "§7.5 — مدیر پروژه نیمرخ متقاضی و امتیاز تطابق را کنار هم می‌بیند.\n\nمرتب‌سازی بر اساس امتیاز تطابق است، نه تاریخ: مدیری که بیست درخواست\nدارد باید از بالا بخواند.",
        "operationId": "list_applications_api_v1_projects__project_id__applications_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationOut"
                  },
                  "title": "Response List Applications Api V1 Projects  Project Id  Applications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "فهرست درخواست‌ها",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "description": "§7.5 — چهار شرط بررسی و امتیاز تطابق عکس‌برداری می‌شود.",
        "operationId": "apply_to_project_api_v1_projects__project_id__applications_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "درخواست پیوستن",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}